+++ src/kestrel_jupyter_kernel/kernel.py
@@ -7,6 +7,9 @@ from kestrel.session import Session
 from kestrel_jupyter_kernel.config import LOG_FILE_NAME
 
 
+_logger = logging.getLogger(__name__)
+
+
 class KestrelKernel(Kernel):
     implementation = "kestrel"
     implementation_version = "1.0"
@@ -20,7 +23,7 @@ class KestrelKernel(Kernel):
         self.kestrel_session = Session()
         _set_logging(
             self.kestrel_session.debug_mode,
-            os.path.join(self.kestrel_session.runtime_directory, LOG_FILE_NAME)
+            os.path.join(self.kestrel_session.runtime_directory, LOG_FILE_NAME),
         )
 
     def do_complete(self, code, cursor_pos):
@@ -49,6 +52,7 @@ class KestrelKernel(Kernel):
                 )
 
             except Exception as e:
+                _logger.error("Exception occurred", exc_info=True)
                 self.send_response(
                     self.iopub_socket, "stream", {"name": "stderr", "text": str(e)}
                 )