+++ CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog`_.
 
+Unreleased
+==========
+
+Added
+-----
+
+- Entire call stack in log if exception occurred.
+
 1.0.1 (2021-06-22)
 ==================